feat: adopt kernel schema types
