+++ crates/core/src/table/state_arrow.rs
@@ -149,7 +149,13 @@ impl DeltaTableState {
             .map(
                 |name| -> Result<arrow::datatypes::DataType, DeltaTableError> {
                     let schema = metadata.schema()?;
-                    let field = schema.field_with_name(name)?;
+                    let field =
+                        schema
+                            .field(name)
+                            .ok_or(DeltaTableError::MetadataError(format!(
+                                "Invalid partition column {0}",
+                                name
+                            )))?;
                     Ok(field.data_type().try_into()?)
                 },
             )
@@ -173,11 +179,11 @@ impl DeltaTableState {
                 .map(|name| -> Result<_, DeltaTableError> {
                     let physical_name = self
                         .schema()
-                        .field_with_name(name)
-                        .or(Err(DeltaTableError::MetadataError(format!(
+                        .field(name)
+                        .ok_or(DeltaTableError::MetadataError(format!(
                             "Invalid partition column {0}",
                             name
-                        ))))?
+                        )))?
                         .physical_name()?
                         .to_string();
                     Ok((physical_name, name.as_str()))
@@ -674,7 +680,6 @@ impl<'a> SchemaLeafIterator<'a> {
         SchemaLeafIterator {
             fields_remaining: schema
                 .fields()
-                .iter()
                 .map(|field| (vec![field.name().as_ref()], field.data_type()))
                 .collect(),
         }
