+++ python/Cargo.toml
@@ -15,6 +15,8 @@ crate-type = ["cdylib"]
 doc = false
 
 [dependencies]
+delta_kernel.workspace = true
+
 # arrow
 arrow-schema = { workspace = true, features = ["serde"] }
